+++ src/main/java/dev/willyelton/crystal_tools/client/events/RegisterKeyBindingsEvent.java
@@ -14,6 +14,7 @@ public class RegisterKeyBindingsEvent {
     public static KeyMapping VEIN_MINE;
     public static KeyMapping MODE_SWITCH;
     public static KeyMapping OPEN_BACKPACK;
+    public static KeyMapping TRIGGER_ROCKET;
 
     private static KeyMapping createBinding(String name, int key, RegisterKeyMappingsEvent e) {
         KeyMapping keyBinding = new KeyMapping(getKey(name), key, getKey("category"));
@@ -27,6 +28,7 @@ public class RegisterKeyBindingsEvent {
         VEIN_MINE = createBinding("vein_mine", GLFW.GLFW_KEY_GRAVE_ACCENT, e);
         MODE_SWITCH = createBinding("mode_switch", GLFW.GLFW_KEY_M, e);
         OPEN_BACKPACK = createBinding("open_backpack", GLFW.GLFW_KEY_B, e);
+        TRIGGER_ROCKET = createBinding("trigger_rocket", GLFW.GLFW_KEY_R, e);
     }
 
     private static String getKey(String name) {