+++ src/main/java/dev/willyelton/crystal_tools/common/network/data/TriggerRocketPayload.java
@@ -1,0 +1,18 @@
+package dev.willyelton.crystal_tools.common.network.data;
+
+import dev.willyelton.crystal_tools.CrystalTools;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import net.minecraft.resources.ResourceLocation;
+
+public record TriggerRocketPayload() implements CustomPacketPayload {
+    public static final TriggerRocketPayload INSTANCE = new TriggerRocketPayload();
+    public static final Type<TriggerRocketPayload> TYPE = new Type<>(ResourceLocation.fromNamespaceAndPath(CrystalTools.MODID, "trigger_rocket"));
+    public static final StreamCodec<RegistryFriendlyByteBuf, TriggerRocketPayload> STREAM_CODEC = StreamCodec.unit(INSTANCE);
+
+    @Override
+    public Type<? extends CustomPacketPayload> type() {
+        return TYPE;
+    }
+}